list result components added
