+++ setup.txt
@@ -1,0 +1,5 @@
+brew install node
+brew install --HEAD watchman
+brew install flow
+npm install -g react-native-cli
+react-native init myProject
